refactor: SettingPage에 대한 컴포넌트 리팩토링 진행
