--- fourtoon-cookie/src/pages/SettingPage/MenuLayout/MenuLayout.styled.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-
-export const styles = StyleSheet.create({
-  menuContainer: {
-    flexDirection: 'column',
-    alignItems: 'flex-start',
-    gap: 20,
-    width: 350,
-  },
-  deleteText: {
-    color: '#F60D0D',
-},
-});
