--- fourtoon-cookie/src/pages/SettingPage/ProfileLayout/ProfileLayout.tsx
@@ -1,39 +1,0 @@
-import { Image, Text, TouchableOpacity, View } from 'react-native';
-import { NavigationProp, useNavigation } from '@react-navigation/native';
-import { RootStackParamList } from '../../../types/routing';
-import { useSelectedCharacterStore } from '../../../hooks/store/selectedCharacter';
-
-import * as S from "./ProfileLayout.styled";
-import { useMember } from '../../../hooks/server/member';
-import { useFunctionWithErrorHandling } from '../../../hooks/error';
-import { useTranslationWithParentName } from '../../../hooks/locale';
-
-const ProfileLayout = () => {
-
-    const { data: member } = useMember();
-
-    const navigation = useNavigation<NavigationProp<RootStackParamList>>();
-    const { selectedCharacter } = useSelectedCharacterStore();
-
-    const { functionWithErrorHandling } = useFunctionWithErrorHandling();
-
-    const t = useTranslationWithParentName('pages.settingPage.profileLayout');
-    const loginT = useTranslationWithParentName('login');
-
-    const handleCharacterChangeButtonPress = functionWithErrorHandling(() => {
-        navigation.navigate('CharacterSelectPage');
-    })
-
-    return (
-        <View style={S.styles.profileContainer}>
-            <Image source={{uri: selectedCharacter?.selectionThumbnailUrl}} style={S.styles.profileImage} />
-            <Text style={S.styles.name}>{member?.name}</Text>
-            <Text style={S.styles.email}>{member ? null : loginT("loginRequired")}</Text>
-            <TouchableOpacity style={S.styles.changeCharacterButton} onPress={handleCharacterChangeButtonPress}>
-                <Text style={S.styles.changeCharacterText}>{t("characterChange")}</Text>
-            </TouchableOpacity>
-      </View>
-    )
-}
-
-export default ProfileLayout;
