--- fourtoon-cookie/src/pages/SettingPage/SettingPage.styled.ts
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    padding: 20,
-  }
-  });
-
